refactor(web): remove redundant add-tool-modal components and related code (#27996)
