+++ web/app/components/base/icons/src/vender/line/others/index.ts
@@ -1,10 +1,7 @@
-export { default as Apps02 } from './Apps02'
 export { default as BubbleX } from './BubbleX'
 export { default as Colors } from './Colors'
 export { default as DragHandle } from './DragHandle'
 export { default as Env } from './Env'
-export { default as Exchange02 } from './Exchange02'
-export { default as FileCode } from './FileCode'
 export { default as GlobalVariable } from './GlobalVariable'
 export { default as Icon3Dots } from './Icon3Dots'
 export { default as LongArrowLeft } from './LongArrowLeft'
