+++ web/i18n/ro-RO/tools.ts
@@ -2,7 +2,6 @@ const translation = {
   title: 'Instrumente',
   createCustomTool: 'Creează Instrument Personalizat',
   type: {
-    all: 'Toate',
     builtIn: 'Incorporat',
     custom: 'Personalizat',
     workflow: 'Flux de lucru',
@@ -20,7 +19,6 @@ const translation = {
     setupModalTitleDescription: 'După configurarea credențialelor, toți membrii din spațiul de lucru pot utiliza acest instrument la orchestrarea aplicațiilor.',
   },
   includeToolNum: '{{num}} instrumente incluse',
-  addTool: 'Adaugă Instrument',
   createTool: {
     title: 'Creează Instrument Personalizat',
     editAction: 'Configurează',
@@ -143,8 +141,6 @@ const translation = {
   addToolModal: {
     added: 'adăugat',
     category: 'categorie',
-    manageInTools: 'Gestionați în Instrumente',
-    add: 'adăuga',
     type: 'tip',
     custom: {
       title: 'Niciun instrument personalizat disponibil',
