+++ web/i18n/uk-UA/tools.ts
@@ -2,7 +2,6 @@ const translation = {
   title: 'Інструменти',
   createCustomTool: 'Створити власний інструмент',
   type: {
-    all: 'Усі',
     builtIn: 'Вбудовані',
     custom: 'Користувацькі',
     workflow: 'Робочий процес',
@@ -20,7 +19,6 @@ const translation = {
     setupModalTitleDescription: 'Після налаштування облікових даних усі члени робочого простору можуть використовувати цей інструмент під час оркестрування програм.',
   },
   includeToolNum: '{{num}} інструмент(ів) включено',
-  addTool: 'Додати інструмент ',
   createTool: {
     title: 'Створити власний інструмент',
     editAction: 'Налаштування',
@@ -142,10 +140,8 @@ const translation = {
   howToGet: 'Як отримати',
   addToolModal: {
     category: 'категорія',
-    add: 'Додати',
     added: 'Додано',
     type: 'тип',
-    manageInTools: 'Керування в інструментах',
     custom: {
       title: 'Немає доступного користувацького інструмента',
       tip: 'Створити користувацький інструмент',
