+++ web/i18n/vi-VN/tools.ts
@@ -2,7 +2,6 @@ const translation = {
   title: 'Công cụ',
   createCustomTool: 'Tạo công cụ tùy chỉnh',
   type: {
-    all: 'Tất cả',
     builtIn: 'Tích hợp sẵn',
     custom: 'Tùy chỉnh',
     workflow: 'Quy trình làm việc',
@@ -20,7 +19,6 @@ const translation = {
     setupModalTitleDescription: 'Sau khi cấu hình thông tin đăng nhập, tất cả thành viên trong không gian làm việc có thể sử dụng công cụ này khi triển khai ứng dụng.',
   },
   includeToolNum: 'Bao gồm {{num}} công cụ',
-  addTool: 'Thêm công cụ',
   createTool: {
     title: 'Tạo công cụ tùy chỉnh',
     editAction: 'Cấu hình',
@@ -142,9 +140,7 @@ const translation = {
   howToGet: 'Cách nhận',
   addToolModal: {
     category: 'loại',
-    manageInTools: 'Quản lý trong Công cụ',
     type: 'kiểu',
-    add: 'thêm',
     added: 'Thêm',
     custom: {
       title: 'Không có công cụ tùy chỉnh nào',
