+++ web/i18n/zh-Hans/tools.ts
@@ -3,7 +3,6 @@ const translation = {
   createCustomTool: '创建自定义工具',
   customToolTip: '了解更多关于 Dify 自定义工具的信息',
   type: {
-    all: '全部',
     builtIn: '工具',
     custom: '自定义',
     workflow: '工作流',
@@ -21,13 +20,10 @@ const translation = {
     setupModalTitleDescription: '配置凭据后，工作区中的所有成员都可以在编排应用程序时使用此工具。',
   },
   includeToolNum: '包含 {{num}} 个 {{action}}',
-  addTool: '添加工具',
   addToolModal: {
     type: '类型',
     category: '类别',
-    add: '添加',
     added: '已添加',
-    manageInTools: '去工具列表管理',
     custom: {
       title: '没有可用的自定义工具',
       tip: '创建自定义工具',
